requirements-local.txt
-r requirements-test.txt
django-debug-toolbar
django-extensions
pre-commit
pydotplus
sphinx
